clean up extension loading + files on client
